+++ Lead-Server-Source/game/src/cmd_gm.cpp
@@ -208,7 +208,7 @@ bool FindInString(const char * c_pszFind, const char * c_pszIn)
 
 		do
 		{
-			strlcpy(sz, c, MIN(sizeof(sz), static_cast<size_t>((p - c) + 1)));
+			strlcpy(sz, c, MIN(static_cast<int>(sizeof(sz)), static_cast<int>((p - c) + 1)));
 
 			if (!_strnicmp(c_pszFind, sz, strlen(c_pszFind)))
 				return true;
